avoid duplicates and missed updates in watchloop
When a `wait_index` is specified, `watchloop` was incrementing it by `1` in every loop.
This can result in it getting duplicates if the etcd index is advanced in the meantime by updates elsewhere.

Also, when `wait_index` is not specified, there is a likelyhood of missing out on updates that would have happened between calls to `get`.

This fixes both the above issues by resetting `wait_index` in each loop from the `modifiedIndex` of the response received in that loop.
Also few minor fixes in tests.